SVG.jl

Documentation for SVG.jl

Missing docstring.

Missing docstring for Style. Check Documenter's build log for details.

SVG.Svg — Type

Svg

Svg object

Properties

objects - Vector of SVG objects

Constructors

Call either empty or with Vector of existing objects

Svg()

Svg(objs::SvgObject)

SVG.Polyline — Type

Polyline

Polyline object

Properties

xs Vector of X coordinates, one per vertex

ys Vector of Y coordinates, one per vertex

Constructors

Quite the choice.

Construct with separate vectors or vectors of 2d points.

Supply an optional transformation function fx, fy, fxy

Or empty or with pre-allocated vectors of xs/ys

Supply Style using kwarg

Polyline(; style::Style=Inline())
Polyline(npoints::Int; style::Style=Inline())
Polyline(xs::Vector, ys::Vector; style::Style=Inline()) 
Polyline(xs::Vector, ys::Vector, fx::Function, fy::Function; style::Style=Inline()) 
Polyline(xys::Vector; style::Style=Inline()) 
Polyline(xys::Vector, fxy::Function; style::Style=Inline()) 
Polyline(xys::Vector, fx::Function, fy::Function; style::Style=Inline())

SVG.Circle — Type

Circle

Circle object

Properties

x, y, r - position and radius

style = Style object

Radius is invariant under the module supplied scaling operations

SVG.Line — Type

Line

Properties

x1, y1 x2, y2 - Endpoints of the line

style - Style object for this line

Constructor

Line(x1, y1, x2, y2; style::Style=Inline())

Base.write — Function

write(fn::String, svg::Svg, width, height; viewbox="", inhtml=false, digits=2, objwrite_fn=write_objs)
write(io::IO, svg::Svg, width, height; viewbox="", inhtml=false, digits=2, objwrite_fn=write_objs)

write the Svg to a file or IO object.

Arguments

fn - write to filename

io - write to IO Object

svg - the Svg to write

width, height - the width / height of the SVG object to supply

viewbox - specify a particular viewbox: "xmin ymin xmax ymax"

inhtml - Boolean flag whether to wrap the SVG in an HTML document for viewing in a browser.

digits - round all SVG Object co-ordinates to this many digits

objwrite_fn - a callback so the caller can modify the objects as they are written

SVG.bounds — Function

bounds(svg::Svg)	
bounds(svg::SvgObject)

Return the bounding box of the entire Svg or an individual SvgObject

Returns

NamedTuple(xmin, ymin, xmax, ymax)

SVG.bounds_viewbox — Function

bounds_viewbox(svg::Svg)
bounds_viewbox(bounds::NamedTuple)

Create a viewbox string using either the Svg or a set of bounds

SVG.scaled — Function

scaled(s::Svg, w, h)
scaled(s::Svg, fx, fy)
scaled(s::SvgObject, fx, fy)

Scale either the whole Svg or individual Objects

s either the whole Svg or individual objects

fx, fy - functions to scale the positional co-ordinates

w, h - scale the entire Svg up to the smallest of w / h and maintain aspect ratio

SVG.points2circles — Function

points2circles(xs, ys, r::Real)
points2circles(xys, r::Real)
points2circles(xs, ys, rs::Vector{Real})
points2circles(xys, rs::Vector{Real})

Helper function to create a collection of points of a given radius to a collection of Circle objects

Arguments

xs, ys Vectors or x/y corrdinates

xys Vector of 2D points

r Radius

rs Vector of Radii